// src/types/supabase.ts
export type Json =
	| string
	| number
	| boolean
	| null
	| { [key: string]: Json | undefined }
	| Json[]

export type Database = {
	public: {
		Tables: {
// ...
			maps: {
				Row: {
					ar: number
					artist: string
					beatmap_id: number
					beatmapset_id: number
					bpm: number
					created_at: string
					cs: number
					difficulty: string
					dt_rate: number | null
					id: number
					length: string
					mapper: string
					mappool_id: number
					mod: Database['public']['Enums']['mod']
					mod_settings: Json[] | null
					name: string
					od: number
					slot: string
					sr: number
					sub_mod: Database['public']['Enums']['sub_mod'] | null
				}
				Insert: {
					ar: number
					artist: string
					beatmap_id: number
					beatmapset_id: number
					bpm: number
					created_at?: string
					cs: number
					difficulty: string
					dt_rate?: number | null
					id?: number
					length: string
					mapper: string
					mappool_id: number
					mod: Database['public']['Enums']['mod']
					mod_settings?: Json[] | null
					name: string
					od: number
					slot: string
					sr: number
					sub_mod?: Database['public']['Enums']['sub_mod'] | null
				}
				Update: {
					ar?: number
					artist?: string
					beatmap_id?: number
					beatmapset_id?: number
					bpm?: number
					created_at?: string
					cs?: number
					difficulty?: string
					dt_rate?: number | null
					id?: number
					length?: string
					mapper?: string
					mappool_id?: number
					mod?: Database['public']['Enums']['mod']
					mod_settings?: Json[] | null
					name?: string
					od?: number
					slot?: string
					sr?: number
					sub_mod?: Database['public']['Enums']['sub_mod'] | null
				}
				Relationships: [
					{
						foreignKeyName: 'public_maps_mappoolid_fkey'
						columns: ['mappool_id']
						isOneToOne: false
						referencedRelation: 'mappools'
						referencedColumns: ['id']
					}
				]
			}
// ...
		}
// ...
	}
}

type PublicSchema = Database[Extract<keyof Database, 'public'>]

export type Tables<
	PublicTableNameOrOptions extends
		| keyof (PublicSchema['Tables'] & PublicSchema['Views'])
		| { schema: keyof Database },
	TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
		? keyof (Database[PublicTableNameOrOptions['schema']]['Tables'] &
				Database[PublicTableNameOrOptions['schema']]['Views'])
		: never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
	? (Database[PublicTableNameOrOptions['schema']]['Tables'] &
			Database[PublicTableNameOrOptions['schema']]['Views'])[TableName] extends {
			Row: infer R
		}
		? R
		: never
	: PublicTableNameOrOptions extends keyof (PublicSchema['Tables'] &
				PublicSchema['Views'])
		? (PublicSchema['Tables'] &
				PublicSchema['Views'])[PublicTableNameOrOptions] extends {
				Row: infer R
			}
			? R
			: never
		: never
// ...
export type Enums<
	PublicEnumNameOrOptions extends
		| keyof PublicSchema['Enums']
		| { schema: keyof Database },
	EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
		? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
		: never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
	? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
	: PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
		? PublicSchema['Enums'][PublicEnumNameOrOptions]
		: never